Initial Upload and update
Began laying out the foundation of this project, I have the countdown timer already going, so if I need to alter it I can. I need to figure out how to animate the card each update and to set the text at the forefront of the "card". I also need to set the stylizing elements.